requirements.txt
accelerate==0.21.0
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
black==23.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
cmake==3.27.2
filelock==3.12.2
frozenlist==1.4.0
fsspec==2023.6.0
huggingface-hub==0.16.4
idna==3.4
Jinja2==3.1.2
lit==16.0.6
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.1
numpy==1.25.2
openai==0.27.8
packaging==23.1
pandas==2.0.3
pathspec==0.11.2
Pillow==10.0.0
platformdirs==3.11.0
psutil==5.9.5
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.2
scipy==1.11.2
sentencepiece==0.1.99
six==1.16.0
sympy==1.12
tiktoken==0.5.1
tokenizers==0.13.3
tomli==2.0.1
tqdm==4.66.1
transformers==4.31.0
triton==2.0.0
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
yarl==1.9.2
scikit-learn